requirements.txt
Django==1.5
South
mysql-python
django-bootstrap3
sendwithus
django-tastypie
